+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/validation_attributes.php
@@ -8,7 +8,7 @@ $container->loadFromExtension('framework', [
     'secret' => 's3cr3t',
     'validation' => [
         'enabled' => true,
-        'enable_annotations' => true,
+        'enable_attributes' => true,
         'email_validation_mode' => 'html5',
     ],
 ]);