+++ src/Symfony/Component/Validator/ValidatorBuilder.php
@@ -17,6 +17,7 @@ use Symfony\Component\Validator\Exception\ValidatorException;
 use Symfony\Component\Validator\Mapping\Factory\LazyLoadingMetadataFactory;
 use Symfony\Component\Validator\Mapping\Factory\MetadataFactoryInterface;
 use Symfony\Component\Validator\Mapping\Loader\AnnotationLoader;
+use Symfony\Component\Validator\Mapping\Loader\AttributeLoader;
 use Symfony\Component\Validator\Mapping\Loader\LoaderChain;
 use Symfony\Component\Validator\Mapping\Loader\LoaderInterface;
 use Symfony\Component\Validator\Mapping\Loader\StaticMethodLoader;
@@ -43,7 +44,7 @@ class ValidatorBuilder
     private array $xmlMappings = [];
     private array $yamlMappings = [];
     private array $methodMappings = [];
-    private bool $enableAnnotationMapping = false;
+    private bool $enableAttributeMapping = false;
     private ?MetadataFactoryInterface $metadataFactory = null;
     private ConstraintValidatorFactoryInterface $validatorFactory;
     private ?CacheItemPoolInterface $mappingCache = null;
@@ -181,29 +182,53 @@ class ValidatorBuilder
     }
 
     /**
-     * Enables annotation and attribute based constraint mapping.
+     * @deprecated since Symfony 6.4, use "enableAttributeMapping()" instead.
      *
      * @return $this
      */
     public function enableAnnotationMapping(): static
     {
+        trigger_deprecation('symfony/validator', '6.4', 'Method "%s()" is deprecated, use "enableAttributeMapping()" instead.', __METHOD__);
+
+        return $this->enableAttributeMapping();
+    }
+
+    /**
+     * Enables attribute-based constraint mapping.
+     *
+     * @return $this
+     */
+    public function enableAttributeMapping(): static
+    {
         if (null !== $this->metadataFactory) {
-            throw new ValidatorException('You cannot enable annotation mapping after setting a custom metadata factory. Configure your metadata factory instead.');
+            throw new ValidatorException('You cannot enable attribute mapping after setting a custom metadata factory. Configure your metadata factory instead.');
         }
 
-        $this->enableAnnotationMapping = true;
+        $this->enableAttributeMapping = true;
 
         return $this;
     }
 
     /**
-     * Disables annotation and attribute based constraint mapping.
+     * @deprecated since Symfony 6.4, use "disableAttributeMapping()" instead
      *
      * @return $this
      */
     public function disableAnnotationMapping(): static
     {
-        $this->enableAnnotationMapping = false;
+        trigger_deprecation('symfony/validator', '6.4', 'Method "%s()" is deprecated, use "disableAttributeMapping()" instead.', __METHOD__);
+
+        return $this->disableAttributeMapping();
+    }
+
+    /**
+     * Disables attribute-based constraint mapping.
+     *
+     * @return $this
+     */
+    public function disableAttributeMapping(): static
+    {
+        $this->enableAttributeMapping = false;
 
         return $this;
     }
@@ -215,7 +240,7 @@ class ValidatorBuilder
      */
     public function setMetadataFactory(MetadataFactoryInterface $metadataFactory): static
     {
-        if (\count($this->xmlMappings) > 0 || \count($this->yamlMappings) > 0 || \count($this->methodMappings) > 0 || $this->enableAnnotationMapping) {
+        if (\count($this->xmlMappings) > 0 || \count($this->yamlMappings) > 0 || \count($this->methodMappings) > 0 || $this->enableAttributeMapping) {
             throw new ValidatorException('You cannot set a custom metadata factory after adding custom mappings. You should do either of both.');
         }
 
@@ -309,8 +334,8 @@ class ValidatorBuilder
             $loaders[] = new StaticMethodLoader($methodName);
         }
 
-        if ($this->enableAnnotationMapping) {
-            $loaders[] = new AnnotationLoader();
+        if ($this->enableAttributeMapping) {
+            $loaders[] = new AttributeLoader();
         }
 
         return array_merge($loaders, $this->loaders);